Move logic to read table up front out of _guess
At the beginning of _guess there was a code block that tries to read the
`table` input as a file or readable file object.  This was moved up one
level from _guess so that if there is just one guess_kwargs then the same
already-read `table` is used for the non-guess reading path.